complete tests, minor change in behavior
The extension now overrules all previous data - this will keep all other extensions consistent, but might be incompatible with them unless this one is run first.

The tests now cover the complete listener with a huge number of combinations (577). This should ensure that we put no security hole into the edit permissions by accident. Unusable combinations of rights and inconsistent rights and event data are included to ensure defined behavior.